pyproject.toml
[project]
name = "gewaechshaus"
version = "0.1.0"
description = "Full backend including sensor data client and API to fully automate my greenhouse."
authors = [
    {name = "Jannik Kalkmann"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "adafruit-circuitpython-dht (>=4.0.9,<5.0.0)",
    "pydantic (>=2.12.0,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "rpi-gpio (>=0.7.1,<0.8.0)",
    "w1thermsensor (>=2.3.0,<3.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "rplcd (>=1.4.0,<2.0.0)",
    "smbus2 (>=0.5.0,<0.6.0)",
    "mqtt (>=0.0.1,<0.0.2)",
    "paho-mqtt (>=2.1.0,<3.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
